Abstract a Status component for username in navbar
