+++ src/styles/_navigationBar.scss
@@ -33,6 +33,10 @@
   .pt-navbar-heading {
     text-transform: uppercase;
   }
+
+  .navbar-username {
+    margin-left: 0.4rem;
+  }
 }
 
 /*
